+++ src/onegov.election_day/onegov/election_day/views/election/main.py
@@ -103,7 +103,6 @@ def view_election_json(self, request):
         'data': {
             'json': request.link(self, 'data-json'),
             'csv': request.link(self, 'data-csv'),
-            'xlsx': request.link(self, 'data-xlsx'),
         }
     }