+++ src/onegov.election_day/onegov/election_day/views/election_compound/main.py
@@ -73,7 +73,6 @@ def view_election_compound_json(self, request):
         'data': {
             'json': request.link(self, 'data-json'),
             'csv': request.link(self, 'data-csv'),
-            'xlsx': request.link(self, 'data-xlsx'),
         }
     }